Ainda organizando pastas e renomeando

+++ Iniciante/uri_1155-Sequência S.c
@@ -8,7 +8,7 @@ int main(){
 	int a;
 
     for(a = 2; a < 101; a++){
-		s += (double)1/a;
+		s += 1.0/a;
     }
 
     printf("%.2lf\n", s);
